refactor(post-manager-service): rebranding service to post-manager-service and dialosoft
