+++ src/post-manager-service/.gitignore
