+++ src/post-manager-service/gradle.properties
@@ -1,0 +1,1 @@
+systemProp.https.protocols=TLSv1.2
