+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/config/OpenApiConfig.java
@@ -1,0 +1,34 @@
+package com.dialosoft.postmanager.config;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
+import io.swagger.v3.oas.annotations.info.Info;
+import io.swagger.v3.oas.annotations.security.SecurityScheme;
+import io.swagger.v3.oas.annotations.security.SecuritySchemes;
+import io.swagger.v3.oas.annotations.servers.Server;
+import io.swagger.v3.oas.models.OpenAPI;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.List;
+
+@Configuration
+@OpenAPIDefinition(
+        info = @Info(title = "Post Manage Service API", version = "1.0", description = "API for managing posts"),
+        servers = @Server(url = "/post-manager-service")
+)
+@SecuritySchemes({
+        @SecurityScheme(name = "bearerAuth", type = SecuritySchemeType.HTTP, scheme = "bearer", bearerFormat = "JWT")
+})
+public class OpenApiConfig {
+    @Bean
+    public OpenAPI userOpenAPI(
+            @Value("${openapi.service.title}") String serviceTitle,
+            @Value("${openapi.service.version}") String serviceVersion,
+            @Value("${openapi.service.url}") String url) {
+        return new OpenAPI()
+                .servers(List.of(new io.swagger.v3.oas.models.servers.Server().url(url)))
+                .info(new io.swagger.v3.oas.models.info.Info().title(serviceTitle).version(serviceVersion));
+    }
+}
