+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/models/Comments.java
@@ -1,0 +1,15 @@
+package com.dialosoft.postmanager.models;
+
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@Setter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class Comments {
+    String username;
+    String content;
+}
