+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/models/entities/ExtraInfoEntity.java
@@ -1,0 +1,4 @@
+package com.dialosoft.postmanager.models.entities;
+
+public class ExtraInfoEntity {
+}
