+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/models/web/request/CreateCommentRequest.java
@@ -1,0 +1,15 @@
+package com.dialosoft.postmanager.models.web.request;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+@Getter
+@Setter
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class CreateCommentRequest {
+    String username;
+    String postId;
+    String content;
+}
