+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/repository/CommentsRepository.java
@@ -1,0 +1,9 @@
+package com.dialosoft.postmanager.repository;
+
+import com.dialosoft.postmanager.models.entities.CommentsEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface CommentsRepository extends JpaRepository<CommentsEntity, UUID> {
+}
