+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/repository/PostManagerRepository.java
@@ -1,0 +1,13 @@
+package com.dialosoft.postmanager.repository;
+
+import com.dialosoft.postmanager.models.entities.PostEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+public interface PostManagerRepository extends JpaRepository<PostEntity, UUID> {
+    // remember using optinal for the methods of this repository
+
+    Optional<List<PostEntity>> findByUsername(String username);
+}
