+++ src/post-manager-service/src/main/java/com/dialosoft/postmanager/services/InteractionsPostService.java
@@ -1,0 +1,7 @@
+package com.dialosoft.postmanager.services;
+
+import com.dialosoft.postmanager.models.web.request.CreateCommentRequest;
+
+public interface InteractionsPostService {
+    void createComment(CreateCommentRequest request);
+}
