+++ src/post-manager-service/src/main/resources/application-docker.yml
@@ -1,0 +1,13 @@
+
+
+eureka:
+  instance:
+    instanceId: ${spring.application.name}:${server.port}
+  client:
+    serviceUrl:
+      defaultZone: ${EUREKA_CLIENT_SERVICEURL_DEFAULTZONE:http://registry-service:8761/eureka/}
+
+openapi:
+  service:
+    # Indicate gateway-service url to enable openapi from there
+    url: http://gateway-service:8080
