+++ src/post-manager-service/src/main/resources/application.yml
@@ -1,0 +1,46 @@
+spring:
+  application:
+    name: post-manager-microservice
+  datasource:
+    url: jdbc:postgresql://${DATASOURCE_HOST}:${DATASOURCE_PORT}/${DATASOURCE_DB}
+    username: ${DATASOURCE_USERNAME}
+    password: ${DATASOURCE_PASSWORD}
+    driver-class-name: org.postgresql.Driver
+  jpa:
+    show-sql: true
+    hibernate:
+      ddl-auto: update
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
+
+logging:
+  level:
+    org:
+      springframework:
+        web: DEBUG
+        servlet: DEBUG
+
+server:
+  port: 8084
+
+# Eureka client config - Registry Service
+eureka:
+  client:
+    fetchRegistry: true
+    registryFetchIntervalSeconds: 5
+
+# OPENAPI
+openapi:
+  service:
+    title: panel-manager-microservice documentation
+    version: 1.0.0
+
+# SWAGGER
+springdoc:
+  api-docs:
+    enabled: true
+    path: /post-manager-service/v3/api-docs
+  swagger-ui:
+    enabled: true
+    path: /post-manager-service/swagger-ui.html
