+++ src/post-manager-service/src/test/java/com/dialosoft/postmanager/PostManagerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.dialosoft.postmanager;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PostManagerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
