--- src/post-manager/src/main/java/com/biznetbb/postmanager/PostManagerApplication.java
@@ -1,15 +1,0 @@
-package com.biznetbb.postmanager;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-
-//@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
-@SpringBootApplication()
-public class PostManagerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PostManagerApplication.class, args);
-	}
-
-}
