--- src/post-manager/src/main/java/com/biznetbb/postmanager/models/Comments.java
@@ -1,15 +1,0 @@
-package com.biznetbb.postmanager.models;
-
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class Comments {
-    String username;
-    String content;
-}
