--- src/post-manager/src/main/java/com/biznetbb/postmanager/models/entities/ExtraInfoEntity.java
@@ -1,4 +1,0 @@
-package com.biznetbb.postmanager.models.entities;
-
-public class ExtraInfoEntity {
-}
