--- src/post-manager/src/main/java/com/biznetbb/postmanager/models/web/request/CreateCommentRequest.java
@@ -1,15 +1,0 @@
-package com.biznetbb.postmanager.models.web.request;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.FieldDefaults;
-
-@Getter
-@Setter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class CreateCommentRequest {
-    String username;
-    String postId;
-    String content;
-}
