--- src/post-manager/src/main/java/com/biznetbb/postmanager/models/web/response/PostManagerResponse.java
@@ -1,19 +1,0 @@
-package com.biznetbb.postmanager.models.web.response;
-
-import com.biznetbb.postmanager.models.Comments;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.FieldDefaults;
-
-import java.util.List;
-
-@Getter
-@Setter
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class PostManagerResponse {
-    String postOwner;
-    String content;
-    String image;
-    List<Comments> comments;
-}
