--- src/post-manager/src/main/java/com/biznetbb/postmanager/services/InteractionsPostService.java
@@ -1,7 +1,0 @@
-package com.biznetbb.postmanager.services;
-
-import com.biznetbb.postmanager.models.web.request.CreateCommentRequest;
-
-public interface InteractionsPostService {
-    void createComment(CreateCommentRequest request);
-}
