--- src/post-manager/src/main/resources/application-dev.yml
@@ -1,14 +1,0 @@
-spring:
-  application:
-    name: post-manager
-  datasource:
-    url: ${POSTGRES_URL:jdbc:postgresql://localhost:5432/test-application} # Preconfigured for postgres 5433, change it for your needs
-    username: ${POSTGRES_USER:postgres}
-    password: ${POSTGRES_PASSWORD:root} # Example default database password
-  jpa:
-    properties:
-      hibernate:
-        default_schema: post-manager
-        dialect: org.hibernate.dialect.PostgreSQLDialect
-    hibernate:
-      ddl-auto: update
