--- src/post-manager/src/test/java/com/biznetbb/postmanager/PostManagerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.biznetbb.postmanager;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PostManagerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
